import script

+++ install/standalone/import.sh
@@ -2,8 +2,9 @@
 
 CMD=`basename $0`
 APPDIR=`dirname $0`
+KEEP=0
 
-while getopts ":ha:" opt; do
+while getopts ":hka:" opt; do
   case $opt in
     h)
 	echo "
@@ -14,11 +15,15 @@ Imports lexicon data for Swedish, Norwegian, US English and a small test file fo
 Options:
   -h help
   -a appdir (default: this script's folder)
+  -k keep lexdata files after import
 " >&2
 	exit 1
       ;;
     a)
         APPDIR=$OPTARG
+      ;;
+    k)
+        KEEP=1
       ;;
     \?)
       echo "Invalid option: -$OPTARG" >&2
@@ -44,6 +49,7 @@ mkdir -p $APPDIR || exit 1
 
 if [ -d "$APPDIR/lexdata" ]; then
     cd $APPDIR/lexdata && git pull && cd -
+    KEEP=1
     else
 	git clone https://github.com/stts-se/lexdata.git $APPDIR/lexdata
 fi
@@ -99,6 +105,13 @@ else
     exit 1
 fi
 
-rm -fr $APPDIR/lexdata
+if [ $KEEP -eq 0 ]; then
+    echo "[$CMD] Clearing lexdata cache" >&2
+    rm -fr $APPDIR/lexdata
+else
+    echo "[$CMD] Keeping lexdata cache" >&2  
+fi
+
+echo "[$CMD] Done. BYE!" >&2
